Extract LoRAConfig and DeviceConfig
